Refactoring: files structure rearranged
