+++ server/src/config/socket.ts
@@ -1,0 +1,33 @@
+import { Server } from "socket.io";
+import http from "http";
+
+
+export const userSocketMap: Record<string, string> = {};
+export let io: Server;
+
+export function setupSocket(server: http.Server) {
+  io = new Server(server, {
+    cors: {
+      origin: process.env.FRONTEND_URL,
+      credentials: true,
+    },
+  });
+
+  io.on("connection", (socket) => {
+    const userId = socket.handshake.query.userId as string | undefined;
+    console.log("User connected:", userId);
+
+    if (userId) {
+      userSocketMap[userId] = socket.id;
+      io.emit("getOnlineUsers", Object.keys(userSocketMap));
+    }
+
+    socket.on("disconnect", () => {
+      console.log("User disconnected:", userId);
+      if (userId) {
+        delete userSocketMap[userId];
+        io.emit("getOnlineUsers", Object.keys(userSocketMap));
+      }
+    });
+  });
+}
